+++ web/app/pages/style-guide/buttons.page.js
@@ -2,44 +2,56 @@ import React from 'react';
 
 import CodeExample from '../../react/components/code-example';
 
-import StylesExample from './assets/examples/buttons/styles';
-import PillExample from './assets/examples/buttons/pill';
-import ThinExample from './assets/examples/buttons/thin';
+import InteractiveExample from './assets/examples/buttons/interactive';
 
 import { readFileSync } from 'fs';
 import { join } from 'path';
 
-const stylesExampleContent = readFileSync(join(__dirname, '/assets/examples/buttons/styles.js'), 'utf8');
-const pillExampleContent = readFileSync(join(__dirname, '/assets/examples/buttons/pill.js'), 'utf8');
-const thinExampleContent = readFileSync(join(__dirname, '/assets/examples/buttons/thin.js'), 'utf8');
+const interactiveExampleContent = readFileSync(join(__dirname, '/assets/examples/buttons/interactive.js'), 'utf8');
 
 class ButtonsPage extends React.Component {
   constructor(props) {
     super(props);
     this.state = {};
+
+    this.onChangeInput = this.onChangeInput.bind(this);
+    this.onBlurInput = this.onBlurInput.bind(this);
+  }
+
+  onChangeInput(event) {
+    var formName = event.target.getAttribute('data-form-name');
+    var fieldName = event.target.getAttribute('data-form-field');
+    var value = getInputValueFromEvent(event);
+
+    var newFormData = onChangeInputStateUpdater.call(this, this.props[formName], {
+      fieldName: fieldName,
+      value: value,
+      markAsDirty: event.target.getAttribute('type') === 'checkbox' || event.target.getAttribute('type') === 'radio'
+    });
+
+    this.props.dispatch(buttonExampleFormActions.set(newFormData));
+  }
+
+  onBlurInput(event) {
+    var formName = event.target.getAttribute('data-form-name');
+    var fieldName = event.target.getAttribute('data-form-field');
+
+    var newFormData = onBlurInputStateUpdater.call(this, this.props[formName], {
+      fieldName: fieldName
+    });
+
+    this.props.dispatch(buttonExampleFormActions.set(newFormData));
   }
 
   render() {
     return (
       <div className="p-style-guide-buttons">
         <h1>Buttons</h1>
-        <h2>Styles</h2>
-        <p>The buttons provide a number of different stylistic options</p>
+        <h2>Interactive Example</h2>
+        <p>This interactive example allows you to play around with all the options of the button to see how everything works.</p>
         <CodeExample
-          exampleComponent={StylesExample}
-          codeContent={stylesExampleContent}
-        />
-        <h2>Pills</h2>
-        <p>Button can be in the form of a pill</p>
-        <CodeExample
-          exampleComponent={PillExample}
-          codeContent={pillExampleContent}
-        />
-        <h2>Thin</h2>
-        <p>Buttons can be inverted with the thin form</p>
-        <CodeExample
-          exampleComponent={ThinExample}
-          codeContent={thinExampleContent}
+          exampleComponent={InteractiveExample}
+          codeContent={interactiveExampleContent}
         />
       </div>
     );