+++ web/app/store.js
@@ -10,12 +10,14 @@ import preventDoubleClick from './stores/prevent-double-click/prevent-double-cli
 import menu from './stores/menu/menu.reducer';
 import applicationNotifications from '../../src/stores/application-notifications/application-notifications.reducer';
 import fileUploadForm from './stores/file-upload-form/file-upload-form.reducer';
+import buttonExampleForm from './stores/button-example-form/button-example-form.reducer';
 
 let myReducers = {
   preventDoubleClick,
   menu,
   applicationNotifications,
-  fileUploadForm
+  fileUploadForm,
+  buttonExampleForm
 };
 
 let reducers = {};