+++ web/app/stores/button-example-form/button-example-form.constants.js
@@ -1,2 +1,2 @@
-export const SET = 'FILE_UPLOAD_FORM::SET';
+export const SET = 'BUTTON_EXAMPLE_FORM::SET';
 export const CLEAR = 'FILE_UPLOAD_FORM::RESET';